requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
pytest
httpx
python-multipart
pinecone
openai
cryptography
llama-cloud-services
llama-index
nest-asyncio
tiktoken
aiofiles
aiolimiter
backoff
aiohttp
asyncio
async-lru
async-timeout
PyPDF2
python-docx
pytest-asyncio